Fix types

+++ main.cpp
@@ -6,8 +6,8 @@ int main(int argc, char const *argv[]) {
     const int S = atoi(argv[1]);
     const int I = atoi(argv[2]);
     const int R = atoi(argv[3]);
-    const double beta = std::stod(argv[4]);
-    const double gamma = std::stod(argv[5]);
+    const float beta = std::stof(argv[4]);
+    const float gamma = std::stof(argv[5]);
     const int t = atoi(argv[6]);
     consoleInterface.start(S, I, R, beta, gamma, t);
   } else {
